refactor(assistant): consume shared workspace from @datum-cloud/datum-ui 1.5.0
