--- app/features/assistant-workspace/components/composer/index.ts
@@ -1,3 +1,0 @@
-export * from './prompt-card';
-export * from './model-selector';
-export * from './equalizer';
