--- app/features/assistant-workspace/components/conversation/index.ts
@@ -1,3 +1,0 @@
-export * from './conversation';
-export * from './turn-rail';
-export * from './scroll-to-bottom-button';
