--- app/features/assistant-workspace/components/message/index.ts
@@ -1,4 +1,0 @@
-export * from './assistant-message';
-export * from './user-message';
-export * from './thinking-block';
-export * from './loading-dots';
