--- app/features/assistant-workspace/components/message/loading-dots.tsx
@@ -1,12 +1,0 @@
-import { cn } from '@datum-cloud/datum-ui/utils';
-
-/** Three bouncing dots — the assistant's "thinking" indicator. */
-export function LoadingDots({ className }: { className?: string }) {
-  return (
-    <div className={cn('flex gap-1', className)}>
-      <span className="bg-muted-foreground/50 h-1.5 w-1.5 animate-bounce rounded-full [animation-delay:-0.3s]" />
-      <span className="bg-muted-foreground/50 h-1.5 w-1.5 animate-bounce rounded-full [animation-delay:-0.15s]" />
-      <span className="bg-muted-foreground/50 h-1.5 w-1.5 animate-bounce rounded-full" />
-    </div>
-  );
-}
