--- app/features/assistant-workspace/components/sidebar/index.ts
@@ -1,2 +1,0 @@
-export * from './chat-rail';
-export * from './history-panel';
