+++ app/features/assistant-workspace/hooks/index.ts
@@ -1,3 +1,2 @@
 export * from './use-chat-logic';
 export * from './use-speech-input';
-export * from './use-turn-rail';
