+++ app/features/assistant-workspace/index.ts
@@ -1,4 +1,4 @@
-export { AssistantWorkspace } from './components/assistant-workspace';
-export { AssistantConfigProvider, useAssistantConfig } from './assistant-config';
+export { AssistantWorkspace } from './assistant-workspace';
 export { STAFF_ASSISTANT_CONFIG } from './staff-config';
 export type { AssistantConfig } from './types';
+export { AssistantConfigProvider, useAssistantConfig } from '@datum-cloud/datum-ui/assistant';
