+++ app/features/assistant-workspace/lib/index.ts
@@ -1,2 +1,1 @@
 export * from './chat-storage';
-export * from './sanitize-html';
