+++ app/features/assistant-workspace/types.ts
@@ -1,61 +1,23 @@
+import type { EffortId } from '@datum-cloud/datum-ui/assistant';
 import type { UIMessage } from 'ai';
-import type { ReactNode } from 'react';
 
-export type EffortId = 'low' | 'medium' | 'high';
-
-export interface ModelOption {
-  /** Stable id sent to the server; must match the allowlist in server/routes/assistant.ts. */
-  id: string;
-  label: string;
-}
-
-export interface EffortOption {
-  id: EffortId;
-  label: string;
-}
-
-/** One submitted prompt — the unit tracked by the right-edge turn rail. */
-export interface Turn {
-  id: string;
-  text: string;
-}
-
-/** Props passed to a `renderLink` implementation (a link in the assistant's markdown). */
-export interface LinkRenderProps {
-  href?: string;
-  children?: ReactNode;
-}
-
-export interface ModelSelectorConfig {
-  models: ModelOption[];
-  efforts: EffortOption[];
-  defaultModelId: string;
-  defaultEffortId: EffortId;
-}
+// Shared assistant types now live in datum-ui; re-export them so the local
+// `../types` imports keep resolving. StoredChat below is staff-specific.
+export type {
+  AssistantConfig,
+  EffortId,
+  EffortOption,
+  LinkRenderProps,
+  ModelOption,
+  ModelSelectorConfig,
+  Turn,
+} from '@datum-cloud/datum-ui/assistant';
 
 /**
- * Everything that differs between host apps (staff vs cloud). Supplied once at
- * the workspace root and read by leaf components via `useAssistantConfig()`, so
- * the presentational pieces stay generic and shell-agnostic.
+ * A locally-persisted chat (localStorage). Staff-specific superset of datum-ui's
+ * `ChatSummary` — adds the fields needed to fully restore a conversation
+ * (per-message HTML, selected model/effort, timestamps).
  */
-export interface AssistantConfig {
-  /** Greeting line on the empty state, e.g. "Hey there, Jacob". */
-  greeting: (name?: string) => string;
-  /** Starter prompts on the empty state. */
-  suggestions: string[];
-  /** Whether to render the assistant's reasoning ("thinking") blocks. */
-  showReasoning: boolean;
-  /** Model + effort picker config, or `false` to hide the control. */
-  modelSelector: ModelSelectorConfig | false;
-  /** Tool name → progress label ("Searching users…"). Host-specific tool set. */
-  toolLabels: Record<string, string>;
-  /** Renders links inside the assistant's markdown (internal routes, external, host-specific icons). */
-  renderLink: (props: LinkRenderProps) => ReactNode;
-  /** Override the per-message action toolbar (defaults to copy + download). */
-  messageActions?: (msg: UIMessage) => ReactNode;
-}
-
-/** A locally-persisted chat (localStorage). */
 export interface StoredChat {
   id: string;
   title: string;
