+++ package.json
@@ -30,7 +30,7 @@
     "@ai-sdk/anthropic": "^4.0.0",
     "@ai-sdk/react": "^4.0.0",
     "@datum-cloud/activity-ui": "0.5.1",
-    "@datum-cloud/datum-ui": "1.3.1",
+    "@datum-cloud/datum-ui": "1.5.0",
     "@epic-web/client-hints": "^1.3.5",
     "@gqlts/runtime": "3.4.2",
     "@hono/prometheus": "^1.0.1",
